# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareConnect API Documentation",
        "description": "API documentation for CareConnect Healthcare Management System",
        "version": "1.0.0",
        "contact": {
            "email": "blissbonheur1@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server V1"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object"
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Success message"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": ["patient", "doctor", "admin"]
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/doctor/create": {
            "post": {
                "tags": ["Doctors"],
                "summary": "Create a Doctor Account",
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["user", "licenseNumber", "specialization", "yearsOfExperience", "hospital"],
                                "properties": {
                                    "user": {
                                        "type": "string",
                                        "description": "ID of the user account"
                                    },
                                    "licenseNumber": {
                                        "type": "string",
                                        "description": "Doctor's medical license number"
                                    },
                                    "specialization": {
                                        "type": "string",
                                        "description": "Doctor's medical specialization"
                                    },
                                    "yearsOfExperience": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "description": "Number of years of medical experience"
                                    },
                                    "hospital": {
                                        "type": "string",
                                        "description": "Name of the hospital or clinic"
                                    },
                                    "availableSlots": {
                                        "type": "array",
                                        "description": "Doctor's available time slots",
                                        "items": {
                                            "type": "object",
                                            "required": ["date", "from", "to"],
                                            "properties": {
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "description": "Date of availability (YYYY-MM-DD)"
                                                },
                                                "from": {
                                                    "type": "string",
                                                    "description": "Start time of availability"
                                                },
                                                "to": {
                                                    "type": "string",
                                                    "description": "End time of availability"
                                                }
                                            }
                                        }
                                    },
                                    "patients": {
                                        "type": "array",
                                        "description": "List of patient IDs assigned to the doctor",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Doctor account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/doctor/all": {
            "get": {
                "tags": ["Doctors"],
                "summary": "List all Doctors",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all doctors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user": {
                                                        "$ref": "#/components/schemas/User"
                                                    },
                                                    "specialization": {
                                                        "type": "string"
                                                    },
                                                    "licenseNumber": {
                                                        "type": "string"
                                                    },
                                                    "yearsOfExperience": {
                                                        "type": "integer"
                                                    },
                                                    "hospital": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/doctor/delete/{id}": {
            "delete": {
                "tags": ["Doctors"],
                "summary": "Delete a Doctor Account",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Doctor ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}